Moving directory 'portfolio',  split into components.
